Add textplots to plot waveform
Looks like the buffer is not guaranteed to be the "fixed size" amount
but rather deals with it as an upper limit. Empirically it seems that
the buffer requests a refill at roughly 1/4 of it's size. This means if
we want to guarantee a buffer of 1024 to work with we need at least 4
times that amount.