Make hardware display buffer larger
Currently the hardware buffer is small enough that a signal processing
thread told to populate the buffer does not finish before another buffer
is requested. This is handled quite well by the firmware (ALSA I assume)
but does print an ugly warning to the screen. This warning interferes
with the tui output. Enlarging the buffer to 256 seems to mostly get rid
of the warnings.